chore: move around defines for correct handling
