+++ srcs/asm/asm.h
@@ -22,45 +22,11 @@
 # include <ft_printf.h>
 # include <op.h>
 
+#include "tokenizer.h"
+
 typedef uint8_t 			t_byte;
 
-typedef enum e_ret			t_ret;
 typedef enum e_line_type	t_line_type;
-typedef enum e_asm_token	t_asm_token;
-
-
-void	debug_print_mem(const void *addr, size_t size); // TODO: remove
-
-// Errors
-//  0 -  9 Standart errors
-// 10 - 29 Input Erros
-// 30 - 39 Translation Errors
-// 40 - 49 Linking errors
-// 50 - 69 Token errors
-//
-
-enum	e_ret
-{
-	kSuccess = 0,
-	kError = -1,
-	kErrorAlloc = -2,
-	kErrorOpeningFile = -3,
-
-	kInputError = -10,
-	kParseError = -11,
-
-	kTranslationError = -30,
-	kInvalidArgumentCount = -31,
-
-	kLinkingError = -40,
-	kLinkNotFoundError = -41,
-
-	kTokenError = -50,
-
-	kMetaParseError = -60,
-	kMPNameTooLong = -61,
-	kMPCommentTooLong = -62,
-};
 
 enum    e_line_type
 {
@@ -74,27 +40,12 @@ enum    e_line_type
 	kEmptyLine,
 };
 
-enum e_asm_token
-{
-	kTokenNone = 0,
-	kTokenLabel,
-	kTokenInstruction,
-	kTokenRegister,
-	kTokenSeperator,
-	kTokenDir,
-	kTokenDirLabel,
-	kTokenInd,
-	kTokenIndLabel,
-	kTokenUnknown,
-};
+
 
 typedef struct s_bytecode	t_bytecode;
 typedef struct s_asm		t_asm;
 typedef struct s_jump		t_jump;
 typedef struct s_jump		t_marker;
-typedef struct s_place		t_place;
-typedef struct s_tksave		t_tksave;
-typedef struct s_error		t_error;
 
 struct	s_bytecode
 {
@@ -125,24 +76,5 @@ struct	s_jump
 	t_args_type	type;
 };
 
-struct	s_place
-{
-	t_index ln;
-	t_index chr;
-};
-
-struct	s_tksave
-{
-	char		*str;
-	t_asm_token	token;
-	t_place		loc;
-};
-
-struct s_error
-{
-	t_ret		code;
-	t_tksave	*token;
-	t_tksave	rtoken;
-};
 
 #endif
