+++ srcs/asm/debugging.h
@@ -19,5 +19,6 @@ void	print_file(char *file);
 void	print_lines(t_list *lines);
 void	print_tokens(t_list *lines);
 void	print_bc(t_asm *asmblr, size_t size);
+void	debug_print_mem(const void *addr, size_t size);
 
 #endif
